RHDM-1056: Change keyboard shortcut for grid multiline input
For more details see https://issues.jboss.org/browse/RHDM-1056
