+++ uberfire-extensions/uberfire-wires/uberfire-wires-core/uberfire-wires-core-grids/src/main/java/org/uberfire/ext/wires/core/grids/client/widget/grid/keyboard/KeyDownHandlerCommon.java
@@ -34,61 +34,35 @@ public class KeyDownHandlerCommon implements KeyDownHandler {
     protected final GridWidget gridWidget;
     protected final HasSingletonDOMElementResource gridCell;
 
-    private final boolean isTabKeyHandled;
-    private final boolean isEnterKeyHandled;
-    private final boolean isEscapeKeyHandled;
-
     public KeyDownHandlerCommon(final GridLienzoPanel gridPanel,
                                 final GridLayer gridLayer,
                                 final GridWidget gridWidget,
                                 final HasSingletonDOMElementResource gridCell) {
-        this(gridPanel,
-             gridLayer,
-             gridWidget,
-             gridCell,
-             true,
-             true,
-             true);
-    }
-
-    public KeyDownHandlerCommon(final GridLienzoPanel gridPanel,
-                                final GridLayer gridLayer,
-                                final GridWidget gridWidget,
-                                final HasSingletonDOMElementResource gridCell,
-                                final boolean isTabKeyHandled,
-                                final boolean isEnterKeyHandled,
-                                final boolean isEscapeKeyHandled) {
         this.gridPanel = gridPanel;
         this.gridLayer = gridLayer;
         this.gridWidget = gridWidget;
         this.gridCell = gridCell;
-
-        this.isTabKeyHandled = isTabKeyHandled;
-        this.isEnterKeyHandled = isEnterKeyHandled;
-        this.isEscapeKeyHandled = isEscapeKeyHandled;
     }
 
     @Override
     public void onKeyDown(final KeyDownEvent e) {
         final int keyCode = e.getNativeKeyCode();
 
+        if (e.isControlKeyDown() && KeyCodes.KEY_ENTER == keyCode) {
+            return;
+        }
+
         switch (keyCode) {
             case KeyCodes.KEY_TAB:
-                if (isTabKeyHandled) {
-                    flush(e);
-                }
+                flush(e);
                 break;
 
             case KeyCodes.KEY_ENTER:
-                if (isEnterKeyHandled) {
-                    flush(e);
-                }
+                flush(e);
                 break;
 
             case KeyCodes.KEY_ESCAPE:
-                if (isEscapeKeyHandled) {
-                    reset();
-                }
+                reset();
         }
 
         e.stopPropagation();
