--- uberfire-extensions/uberfire-wires/uberfire-wires-core/uberfire-wires-core-grids/src/test/java/org/uberfire/ext/wires/core/grids/client/widget/grid/keyboard/KeyDownHandlerCommonTabKeyTest.java
@@ -1,39 +1,0 @@
-/*
- * Copyright 2019 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.uberfire.ext.wires.core.grids.client.widget.grid.keyboard;
-
-import com.ait.lienzo.test.LienzoMockitoTestRunner;
-import org.junit.runner.RunWith;
-
-@RunWith(LienzoMockitoTestRunner.class)
-public class KeyDownHandlerCommonTabKeyTest extends BaseKeyDownHandlerCommonTest {
-
-    @Override
-    protected boolean isTabKeyHandled() {
-        return true;
-    }
-
-    @Override
-    protected boolean isEnterKeyHandled() {
-        return false;
-    }
-
-    @Override
-    protected boolean isEscapeKeyHandled() {
-        return false;
-    }
-}
